+++ app/signup/page.tsx
@@ -1,4 +1,3 @@
-'use client';
 import {
   Button,
   Checkbox,
